Implement word list dictionary
This dictionary service works based on a list of words that does or does
not contain a given word. The concept here is, of course, that you will
ask if the dictionary has a word and then ask if the dictionary has any
suggestions (but probably only if it came back as 'incorrect.'). Maybe
this isn't all that efficient, but, still. This is what we got at the
moment.

Also, the suggestions I've implemented so far are based only on
Levenshtein distance, which is an absolutely awful way to determine what
would be a better spelling of your word.