Update nuget packages
For some reason, it isn't possible to update `Microsoft.NETCore.App`
because it is defined by the project that depends on it. As a result, we
have one version for the web project and another version for the test
project. All hail whoever manages to get those two to agree with one
another.